# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Aerospike Backup Service",
        "title": "Backup Service REST API Specification",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/v1/backups/full": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Backup"
                ],
                "summary": "Get available full backups.",
                "operationId": "getFullBackups",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Lower bound timestamp filter",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Upper bound timestamp filter",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Full backups by routine",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/model.BackupDetails"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/backups/full/{name}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Backup"
                ],
                "summary": "Get available full backups for routine.",
                "operationId": "getFullBackupsForRoutine",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Backup routine name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Lower bound timestamp filter",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Upper bound timestamp filter",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Full backups for routine",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.BackupDetails"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/backups/incremental": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Backup"
                ],
                "summary": "Get available incremental backups.",
                "operationId": "getIncrementalBackups",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Lower bound timestamp filter",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Upper bound timestamp filter",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Incremental backups by routine",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/model.BackupDetails"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/backups/incremental/{name}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Backup"
                ],
                "summary": "Get incremental backups for routine.",
                "operationId": "getIncrementalBackupsForRoutine",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Backup routine name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Lower bound timestamp filter",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Upper bound timestamp filter",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Incremental backups for routine",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.BackupDetails"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/backups/schedule/{name}": {
            "post": {
                "tags": [
                    "Backup"
                ],
                "summary": "Schedule a full backup once per routine name.",
                "operationId": "scheduleFullBackup",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Backup routine name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Delay interval in milliseconds",
                        "name": "delay",
                        "in": "query"
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.BackupDetails": {
            "description": "BackupDetails contains information about a backup.",
            "type": "object",
            "properties": {
                "byte-count": {
                    "description": "The size of the backup in bytes.",
                    "type": "integer"
                },
                "created": {
                    "description": "The backup time in the ISO 8601 format.",
                    "type": "string"
                },
                "file-count": {
                    "description": "The number of backup files created.",
                    "type": "integer"
                },
                "from": {
                    "description": "The lower time bound of backup entities in the ISO 8601 format (for incremental backups).",
                    "type": "string"
                },
                "key": {
                    "description": "The path to the backup files.",
                    "type": "string"
                },
                "namespace": {
                    "description": "The namespace of a backup.",
                    "type": "string"
                },
                "record-count": {
                    "description": "The total number of records backed up.",
                    "type": "integer"
                },
                "secondary-index-count": {
                    "description": "The number of secondary indexes backed up.",
                    "type": "integer"
                },
                "udf-count": {
                    "description": "The number of UDF files backed up.",
                    "type": "integer"
                }
            }
        },
...
    },
...
}